[BE/POST] API 통신 전 모든 기능 테스트 및 디버깅 완료

+++ Backend/src/main/java/com/ssafy/backend/post/service/CommentServiceImpl.java
@@ -1,8 +1,11 @@
 package com.ssafy.backend.post.service;
 
 
-import com.ssafy.backend.jwt.JwtUtil;
+import com.ssafy.backend.cafe.domain.entity.Cafe;
+import com.ssafy.backend.cafe.repository.CafeRepository;
 import com.ssafy.backend.member.domain.entity.Member;
+import com.ssafy.backend.member.domain.entity.MemberCafeTier;
+import com.ssafy.backend.member.repository.MemberCafeTierRepository;
 import com.ssafy.backend.member.repository.MemberRepository;
 import com.ssafy.backend.post.domain.dto.*;
 import com.ssafy.backend.post.domain.entity.Comment;
@@ -10,8 +13,9 @@ import com.ssafy.backend.post.domain.entity.CommentLike;
 import com.ssafy.backend.post.domain.entity.Post;
 import com.ssafy.backend.post.repository.*;
 import com.ssafy.backend.post.util.PostUtil;
+import com.ssafy.backend.redis.CafeAuth;
+import com.ssafy.backend.redis.CafeAuthRepository;
 import lombok.RequiredArgsConstructor;
-import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
 import org.springframework.stereotype.Service;
 import javax.transaction.Transactional;
@@ -24,45 +28,73 @@ public class CommentServiceImpl implements CommentService {
 
     private final PostRepository postRepository;
     private final PostLikeRepository postLikeRepository;
+    private final CafeRepository cafeRepository;
     private final CommentRepository commentRepository;
     private final CommentLikeRepository commentLikeRepository;
     private final MemberRepository memberRepository;
     private Slice<Comment> comments;
-    private final JwtUtil jwtUtil;
     private final PostUtil postUtil;
-    private Slice<Post> postSlice;
     private Post post;
     private Member member;
-    private Comment comment;
+    private final CafeAuthRepository cafeAuthRepository;
+    private final MemberCafeTierRepository memberCafeTierRepository;
 
     /**
      * 2-1. 댓글 불러오기
      **/
     @Override
-    public List<CommentPagingResponseDto> feedComment(CommentPagingRequestDto requestDto, Pageable pageable) throws Exception {
+    public List<CommentPagingResponseDto> feedComment(CommentPagingRequestDto requestDto) throws Exception {
         // 가지고와야할 값 - 댓글, 대댓글 / 작성시간 /
         postUtil.checkMember();
 
         List<CommentPagingResponseDto> commentResponseList = new ArrayList<>();
-        CommentPagingResponseDto commentPagingResponseDto;
         Long commentId = requestDto.getCommentId();
         Long postId = requestDto.getPostId();
-
-        if(commentId == -1) {
-            comments = commentRepository.findAllByPostId(postId);
-        }else {
-            comments = commentRepository.findAllByIdLessThanAndPostId(commentId, postId);
+        Post post = postRepository.findById(postId).get();
+        Long group;
+        if(commentId == -1L) {
+            group = 0L;
+        }
+        else{
+            group = commentRepository.findById(commentId).get().getGroupNo();
         }
 
-        for (Comment comment : comments) {
-            commentPagingResponseDto = CommentPagingResponseDto.CommentResponseBuilder()
-                    .memberId(comment.getMember().getId())
+        // 2.
+        Set<Long> groupSet = new HashSet<>();
+        List<Comment> commentList = post.getCommentList();
+        if(commentList == null || commentList.isEmpty()) {
+            return null;
+        }
+        for (Comment comment: commentList) {
+            if(comment.getGroupNo() > group) groupSet.add(comment.getGroupNo());
+            if(groupSet.size() == 5) break;
+        }
+        System.out.println(groupSet);
+        if(groupSet.isEmpty() || groupSet == null) { // 불러올 그게 없다.
+            return null;
+        }
+        Slice<Comment> commentSlice = commentRepository.findAllByGroupNoInAndPostId(groupSet, postId);
+        if(commentSlice == null || commentSlice.isEmpty()) {
+            return null;
+        }
+
+
+        for (Comment comment : commentSlice) {
+            CommentPagingResponseDto commentPagingResponseDto = CommentPagingResponseDto.CommentResponseBuilder()
+                    .writerId(comment.getMember().getId())
                     .content(comment.getContent())
                     .createdAt(comment.getCreatedAt())
-                    .commentLike(comment.getCommentLikeList().size())
+                    .commentLikeCnt(comment.getCommentLikeList().size())
                     .build();
-
             commentResponseList.add(commentPagingResponseDto);
+            System.out.println("미인증 유저 댓글쓰기 불러오기 완료");
+            Optional<CafeAuth> cafeAuth = cafeAuthRepository.findById(comment.getMember().getNickname());
+            if(cafeAuth.isPresent()) {
+                Cafe cafe = cafeRepository.findById(cafeAuth.get().getCafeId()).get();
+                MemberCafeTier memberCafeTier = memberCafeTierRepository.findByMemberIdAndCafeId(comment.getMember().getId(),cafe.getId()).get();
+                commentPagingResponseDto.updateCommentVerifiedUser(cafe.getId(),cafe.getName(),memberCafeTier.getExp());
+                System.out.println("인증유저 댓글 업데이트 완료");
+            }
         }
 
         return commentResponseList;
@@ -72,37 +104,44 @@ public class CommentServiceImpl implements CommentService {
      * 2-2. 댓글 쓰기 [테스트 완료 - step_no 랑 group_no 알고리즘만 경희랑 이야기해서 짜면될듯]
      **/
     @Override
-    public void writeComment(CommentWriteRequestDTO commentWriteDto) throws Exception {
+    public int writeComment(CommentWriteRequestDTO commentWriteDto) throws Exception {
         //1. 유저 확인
         CheckedResponseDto checked = postUtil.checkMember();
         long memberId = checked.getMemberId(); // 멤버 아이디를 확인한다.
-
         // 2. 값 확인
-        Long commentMemberId = commentWriteDto.getMemberId();
         Long postId = commentWriteDto.getPostId();
         String content = commentWriteDto.getContent();
-        Long group = commentWriteDto.getGroup();
-        Long step = commentWriteDto.getStep();
-
-        // 3. 값 찾아오기
-        Optional<Member> memberOptional = memberRepository.findById(commentMemberId);
-        member = memberOptional.orElseThrow();
+        Optional<Member> memberOptional = memberRepository.findById(memberId);
+        member = memberOptional.get();
 
         Optional<Post> postOptional = postRepository.findById(postId);
-        post = postOptional.orElseThrow();
+        if(postOptional.isEmpty() || postOptional == null) {
+            System.out.println("댓글쓰기 중 : 글이 사라짐");
+            return 1;
+        }
+        post = postOptional.get();
+
+        // 3. 댓글 대댓글 구분
+        Long groupNo = commentWriteDto.getGroup();
+        if(groupNo == -1L) { // 댓글
+            Optional<Comment> commentOptional = commentRepository.findTopByPostIdOrderByIdDesc(postId);
+            if(commentOptional.isEmpty() || commentOptional == null) {
+                groupNo = 1L;
+            }else groupNo = commentOptional.get().getGroupNo() + 1L;
+        }
 
         // 4. 글 저장하기
-        comment = Comment.CommentWriteBuilder()
-                .group(group)
+        Comment comment = Comment.CommentWriteBuilder()
+                .group(groupNo)
                 .member(member)
                 .post(post)
                 .content(content)
-                .step(step)
                 .build();
 
         System.out.println(comment);
-        // 인증 여부에 따라 글을 쓸수있다 - GeoAuth
+        // 인증 여부에 따라 글을 쓸수있다 - GeoAuth - 따로 필요없음
         commentRepository.save(comment);
+        return 2;
     }
 
     /**
@@ -110,53 +149,77 @@ public class CommentServiceImpl implements CommentService {
      **/
 
     @Override
-    public void updateComment(CommentUpdateRequestDTO commentUpdateDto) throws Exception{
+    public int updateComment(CommentUpdateRequestDTO commentUpdateDto) throws Exception{
         CheckedResponseDto checked = postUtil.checkMember();
-        long memberId = checked.getMemberId(); // 멤버 아이디를 확인한다.
-
         Long commentId = commentUpdateDto.getCommentId();
-        String content = commentUpdateDto.getContent();
 
         Optional<Comment> optionalComment = commentRepository.findById(commentId);
+        Comment comment;
         if(optionalComment == null || optionalComment.isEmpty()) {
-            return;
+            return 1;
         }else {
-            Comment comment = optionalComment.orElseThrow();
-            comment.updateComment(content);
+            comment = optionalComment.get();
         }
 
+        Long writerId = commentRepository.findById(commentId).get().getMember().getId();
+        long memberId = checked.getMemberId(); // 멤버 아이디를 확인한다.
+
+        if(writerId != memberId) {
+            // 글쓴이와 수정하려는 사람이 같지 않으면
+            return 2;
+        }
+        String content = commentUpdateDto.getContent();
+        if(content == null || content.isEmpty()) {
+            // content 가 비어있어!
+            return 3;
+        }
+        comment.updateComment(content);
         commentRepository.save(comment);
+        return 4;
     }
 
     /**
      * 3. 댓글 좋아요
      **/
     @Override
-    public Map.Entry<Boolean, Long> likeComment(Long commentId, Boolean isChecked) throws Exception {
+    public CommentLikeResponseDto likeComment(CommentLikeRequestDto requestDto) throws Exception {
         Boolean responseIsChecked;
 
         //1. 유저 확인
         CheckedResponseDto checked = postUtil.checkMember();
         long memberId = checked.getMemberId(); // 멤버 아이디를 확인한다.
+        boolean isChecked = requestDto.isChecked();
+        Long commentId = requestDto.getCommentId();
 
-        // False 라면 생성
+
+        Optional<Comment> optionalComment = commentRepository.findById(commentId);
+        if(optionalComment.isEmpty() || optionalComment == null) {
+            System.out.println("댓글 좋아요 : 댓글 id 가 잘못됨 null");
+            return null;
+        }
+        Comment comment = optionalComment.get();
 
         if (!isChecked) {
             CommentLike commentLike = CommentLike.CommentLikeBuilder()
                     .comment(comment)
-                    .member(member)
+                    .member(memberRepository.findById(memberId).get())
                     .build();
             commentLikeRepository.save(commentLike);
             responseIsChecked = true;
         }
         // True 라면 삭제
         else {
-            commentLikeRepository.deleteByCommentIdAndMemberId(commentId, memberId);
+            commentLikeRepository.deleteById(commentId);
             responseIsChecked = false;
         }
-        Long count = commentLikeRepository.countCommentLikeByCommentId(commentId);
+        int count = commentLikeRepository.countCommentLikeByCommentId(commentId);
 
-        return new AbstractMap.SimpleEntry<>(responseIsChecked, count);
+        CommentLikeResponseDto commentLikeResponseDto = CommentLikeResponseDto.builder()
+                .commentId(commentId)
+                .likeCount(count)
+                .isChecked(responseIsChecked)
+                .build();
+        return commentLikeResponseDto;
     }
 
     /**
@@ -164,22 +227,23 @@ public class CommentServiceImpl implements CommentService {
      **/
 
     @Override
-    public boolean deleteComment(Long commentId) throws Exception{
-
+    public int deleteComment(Long commentId) throws Exception{
         //1. 유저를 확인한다.
         CheckedResponseDto checked = postUtil.checkMember();
         long memberId = checked.getMemberId(); // 멤버 아이디를 확인한다.
 
         //2. 유저(나) 와 댓글유저가 일치하는지 확인한다.
-
         Optional<Comment> commentOptional = commentRepository.findById(commentId);
-        long commentMemberId = commentOptional.orElseThrow().getMember().getId();
+        if(commentOptional.isEmpty() || commentOptional == null) {
+            return 1;
+        }
+        long commentMemberId = commentOptional.get().getMember().getId();
 
         if(memberId == commentMemberId) { // 글쓴이와 유저가 일치한다면 삭제
             commentRepository.deleteById(commentId);
-            return true;
+            return 2;
         } else { // 유저가 일치하지 않는다면 badgateway
-            return false;
+            return 3;
         }
 
     }
